% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionAPGBdate.R
\name{APGBdate}
\alias{APGBdate}
\title{APGBdate}
\usage{
APGBdate(imgfolder, lookup, field = "OS")
}
\arguments{
\item{imgfolder}{folder path containing the imagery to rename}

\item{lookup}{a lookup shapefile to reference dates}

\item{field}{field within the shapefile containing the dates}
}
\description{
APGBdate
}
