% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionBareThresh.R
\name{barethresh}
\alias{barethresh}
\title{thresholding bare peat based on indices}
\usage{
barethresh(Img.path, out.path, spec.bands = NA, ind.name = NA, c.fun,
  nir = 4, r = 1, g = 2, b = 3, swir = NA, start = 1,
  end = NA)
}
\arguments{
\item{Img.path}{path to the folder with indices values or list of file paths}

\item{out.path}{output folder to write output to}

\item{spec.bands}{the spectral band numbers you wish to include in the thresholding rule}

\item{ind.name}{name of indices, will assume that the file naming convention is ind.name_... and will replace with BARE_ when writing out}

\item{c.fun}{function with which to calculate across the indices stack}

\item{nir}{near infrared band in the multispectral imagery}

\item{r}{red band in the multispectral imagery}

\item{g}{green band in the multispectral imagery}

\item{b}{blue band in the multispectral imagery}

\item{swir}{short wave infrared band in the multispectral imagery}

\item{start}{starting image to process, defaults to the first image}

\item{end}{ending image to process, if NULL will default to the last image in the folder}
}
\description{
thresholding bare peat based on indices
}
