% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionChangeStats.R
\name{change_stats}
\alias{change_stats}
\title{change_stats}
\usage{
change_stats(sitefile, outfolder, polygons, polyid = "id",
  habclass = "a_pred")
}
\arguments{
\item{sitefile}{a text file containing the calculated zonal statistics for a site, output from the 'zonal_stats' function}

\item{outfolder}{folder path of where to save outputs}

\item{polygons}{shapefile of the polygons}

\item{polyid}{field in the polygon shapefile containing a unique identifier}

\item{habclass}{field in the polygon shapefile containing the habitat class}
}
\description{
change_stats
}
