% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionClassProfile.R
\name{ClassProfile}
\alias{ClassProfile}
\title{Extract imagery and indices differences between classes}
\usage{
ClassProfile(trainpath, classname, imagepath, indicespath, tcrs = 27700)
}
\arguments{
\item{trainpath}{filepath the training shapefile with different classes}

\item{classname}{field in the training shapefile denoting where the classes are identified}

\item{imagepath}{path to the imagery rasterfiles, can be a file or folder path. if folder is supplied this will iterate through each file in the folder}

\item{indicespath}{path to where the indices for each image have been generated}

\item{tcrs}{the ESPG code of the coordinate reference system of the training shapefile}
}
\description{
Extract imagery and indices differences between classes
}
\examples{


}
