% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctioncomboAPGB.R
\name{comboAPGB}
\alias{comboAPGB}
\title{#### comboAPGB ####
Function designed to combine RGB imagery with additional NIR band from infrared image. It resamples image b to image a resolution then combines the two}
\usage{
comboAPGB(a.path, b.path, nir = 1, out.path, start = 1, end = NULL)
}
\arguments{
\item{a.path}{path to RGB image folder}

\item{b.path}{path to infrared image folder}

\item{nir}{band in b image which contains the Near-infrared band to join}

\item{out.path}{output folder to save to}

\item{start}{image to start on, will use list.files of the .tif files in folder}

\item{end}{image to end on, will use list.files of the .tif files in folder}
}
\description{
#### comboAPGB ####
Function designed to combine RGB imagery with additional NIR band from infrared image. It resamples image b to image a resolution then combines the two
}
