% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_change_detection_functions.R
\name{create_square_bounding_box_polygon}
\alias{create_square_bounding_box_polygon}
\title{Create square bounding box polygon}
\usage{
create_square_bounding_box_polygon(site_polygon)
}
\arguments{
\item{site_polygon, }{sf object, site polygon or multipolygon}
}
\value{
sf object, square polygon bounding box buffered by 10m
}
\description{
This function creates a square bounding box polygon for a site,
buffered by 10m so that it includes a margin of single raster pixels
}
