% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_change_functions.R
\name{fileEdit}
\alias{fileEdit}
\title{Define function to edit the zonal stats txt files}
\usage{
fileEdit(filePath)
}
\arguments{
\item{filePath}{list of filepaths within the app data folder}
}
\value{
overwrites files with edited csvs
}
\description{
Define function to edit the zonal stats txt files
}
\examples{

}
