% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_change_functions.R
\name{fileEdit2}
\alias{fileEdit2}
\title{Define function to write the habitat names into the zonal stats txt files}
\usage{
fileEdit2(filePath, config)
}
\arguments{
\item{filePath}{list of files in the app data folder}

\item{config}{yaml config file with app data structure}
}
\value{
overwrites files with edited csvs
}
\description{
Define function to write the habitat names into the zonal stats txt files
}
\examples{

}
