% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag_change_functions.R
\name{flag_change}
\alias{flag_change}
\title{Function to flag change using the outputs of the 'change_stats' function}
\usage{
flag_change(files, indices)
}
\arguments{
\item{files}{vector of zonal stats files to iterate through}

\item{indices}{vector of indices names}
}
\value{
csv files flagging change
}
\description{
Function to flag change using the outputs of the 'change_stats' function
}
\examples{

}
