% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_change_detection_functions.R
\name{legend_thumbnail}
\alias{legend_thumbnail}
\title{create legend thumbnail}
\usage{
legend_thumbnail(r, index, breaks, palette, labels, outfolder,
  continuous = F)
}
\arguments{
\item{r}{raster layer}

\item{index}{index name, this will be applied for the output filename}

\item{breaks}{breaks to pass into the tm_raster function, see ?tm_raster for more info.}

\item{palette}{colour palette to pass into the tm_raster function, see ?tm_raster for more info.}

\item{labels}{labels to pass into the tm_raster function, see ?tm_raster for more info.}

\item{outfolder}{output folder to save the images}

\item{continuous}{logical, whether to plot on a continuous scale or fixed categories.}
}
\value{
image png of the legend
}
\description{
create legend thumbnail
}
\examples{

}
