% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionMaskBare.R
\name{mask_bare}
\alias{mask_bare}
\title{Bare Peat Masking}
\usage{
mask_bare(bare.path, out.path, incl.masks, excl.masks, start = 1,
  end = NA)
}
\arguments{
\item{bare.path}{path to the bare peat classified image}

\item{out.path}{path to save the outputs to, a new folder called 'training' will be created}

\item{incl.masks}{list of shapefiles to mask the information to, bare peat areas falling within these polygons with be kept}

\item{excl.masks}{list of shapefiles to mask the information from, bare peat areas falling within these polygons with be excluded}

\item{start}{image number to start on - for iterating, otherwise will start on first image}

\item{end}{image number to end on - for iterating, otherwise will take all}
}
\description{
Bare Peat Masking
}
