% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionMaskPeatSoils.R
\name{MaskPeatSoils}
\alias{MaskPeatSoils}
\title{Title}
\usage{
MaskPeatSoils(img_folder, mask, epsg = 27700, delete = T)
}
\arguments{
\item{img_folder}{folder path containing the imagery to iterate through}

\item{mask}{mask to retain the imagery within}

\item{epsg}{coordinate reference system, EPSG code}

\item{delete}{logical, if you wish to delete those images which do not overlap within the mask layer}
}
\description{
Title
}
