% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcovClassify.R
\name{pcov_classify}
\alias{pcov_classify}
\title{pcov_classify}
\usage{
pcov_classify(img_path, bare_path, out_path, polymask)
}
\arguments{
\item{img_path}{folder path to where the Sentinel-2 iamgery is stored}

\item{bare_path}{folder path to where the classified bare peat rasterlayers are stored}

\item{out_path}{folder path to store the outputs}

\item{polymask}{shapefile of polygons to mask to}
}
\description{
pcov_classify
}
