% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/habitat_change_detection_functions.R
\name{reduce_image_size}
\alias{reduce_image_size}
\title{Reduce image size}
\usage{
reduce_image_size(path_image, image_size = "33\%")
}
\arguments{
\item{path_image}{character, path to full size image}

\item{image_size}{character, percentage size of reduced image}
}
\value{
character, path to reduced size image
}
\description{
This function reduces the size of an image, saving the original
image in an archive folder
}
