% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionRFReg.R
\name{RFReg}
\alias{RFReg}
\title{RFReg}
\usage{
RFReg(training, varpredict, varnames, out.folder, max_tries = 5,
  prop.test = 0.25, nsamp = 1000, resamp = 5, stratified = T,
  fillsamp = F)
}
\arguments{
\item{training}{csv file containing training data}

\item{varpredict}{folder path to the time series variable stacks}

\item{varnames}{names of variable layers}

\item{out.folder}{folder path to save the outputs}

\item{max_tries}{number of model runs}

\item{prop.test}{proportion of data to be held back to evaluate the model performance}

\item{nsamp}{number of points from the training data to attempt to sample}

\item{resamp}{numeric, feeds into the caret::trainControl parameter of the RF model. The number of times  K-fold cross-validation is repeated for tuning parameters for the model.}

\item{stratified}{logical, whether to use stratified sampling to obtain training data.Note: currently sampling is expecting a field called 'bare' for sampling and 'barecat' for stratified sampling.}

\item{fillsamp}{logical, whether to fill remaining sampling points from other categories where one category doesnt have enough points}
}
\description{
RFReg
}
