% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionRunIndices.R
\name{runIndices}
\alias{runIndices}
\title{Run Indices}
\usage{
runIndices(imagepath, outpath, nir, r, b, g, swir = NA, indices,
  nf = T)
}
\arguments{
\item{imagepath}{folder path to the multispectral Sentinel-2 imagery}

\item{outpath}{folder path to save output files}

\item{nir}{near infra-red band}

\item{r}{numeric, red band}

\item{b}{numeric, blue band}

\item{g}{numeric, green band}

\item{swir}{numeric, optional - Short wave infrared band}

\item{indices}{a vector of indices to retrieve e.g. c("EVI", "GLI","GNDVI","RDVI","SAVI","SBL"). The function will accept EVI,GLI,GNDVI,RDVI,SAVI,SBL,RB,RG,Brightness,NBR}

\item{nf}{logical, if you wish to create new folder per image}
}
\description{
Run Indices
}
\examples{

}
