% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionRunS1Indices.R
\name{runS1Indices}
\alias{runS1Indices}
\title{runS1Indices}
\usage{
runS1Indices(imagepath, outpath, vv, vh, indices = c("RVI", "RVIv"),
  nf = F)
}
\arguments{
\item{imagepath}{folder path to the Sentinel-1 imagery}

\item{outpath}{path to save NDVI output files}

\item{vv}{numeric, number of the vv band in the imagery}

\item{vh}{numeric, number of the vh band in the imagery}

\item{indices}{a vector of indices to retrieve, e.g. c('RVI','RVIv')}

\item{nf}{logical, if you wish to create new folder per image}
}
\description{
runS1Indices
}
