% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionFalseColourThumb.R
\name{s1thumbs}
\alias{s1thumbs}
\title{s1thumbs}
\usage{
s1thumbs(s1path, out_folder = paste0(outfolder, "ZonalStats/s2/thumbs/"))
}
\arguments{
\item{s1path}{filepath containing the Sentinel-1 imagery}

\item{out_folder}{filepath to save the outputs}
}
\description{
s1thumbs
}
