% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sat_varstack.R
\name{sat_varstack}
\alias{sat_varstack}
\title{sat_varstack}
\usage{
sat_varstack(varpath, satimg, satbands = NA, indices = NA, vars = NA,
  mask = NA, outpath, stackname = "stack")
}
\arguments{
\item{varpath}{folder path to the variable layers}

\item{satimg}{filepath to Sentinel-2 imagery layer}

\item{satbands}{band required from the sentinel imagery}

\item{indices}{optional, vector of indices layers to use, note will search to retreive these from a folder named Indices, see code for naming convention. expects names of indices not filepaths}

\item{vars}{optional, vector of filepaths for additional variable layers to include in the varstack}

\item{mask}{optional, shapefile to mask to}

\item{outpath}{output folder to write to}

\item{stackname}{name for the output rasterstack}
}
\description{
sat_varstack
}
