% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_extract.R
\name{train_extract}
\alias{train_extract}
\title{#### Extract values from variables using matrices ####
# Iterate through and extract points from folder of variable imagery}
\usage{
train_extract(varstack, varnames, trainr, outfolder, stratified = T)
}
\arguments{
\item{varstack}{folder path to the variable layers}

\item{varnames}{names of variable layers}

\item{trainr}{rasterlayer of training values}

\item{outfolder}{output folder to write to}

\item{stratified}{optional, adds a column to the data which categorises the data based on equal intervals between 0-1}
}
\description{
#### Extract values from variables using matrices ####
# Iterate through and extract points from folder of variable imagery
}
